Fix a few remaining call-sites
There were a few places where `.cast<List<int>>()` calls were
missed in #384 and #385, and likewise for `Datagram.data` in #382.

Note that I had to change one of the tests of `Datagram.data`, since
the existing test was setting invalid values as the data (signed
integers or integers that didn't fit in 8-bytes).

dart-lang/sdk#36900
https://github.com/dart-lang/co19/issues/381
https://github.com/dart-lang/co19/issues/383