Fixes #1913. Update generated tests to work with class-modifiers (#1914)
This PR changes the library `generator.dart` such that a large number of generated libraries change as well.

Most changes are `class` --> `mixin class`, which was needed because the given class is used as a mixin. Some changes are concerned with the use of records (a `SharedOptions` line is inserted, and an import is shown in the diff, both as an added and as a deleted line). The changes to `generator.dart` are concerned with the introduction and management of a shared options. Finally, a template (used to generate code) was adjusted: `interface_compositionality_A02.dart`.