Update static_variable_t02.dart
Add a `static late` case along with the `late static` case. The point is that `static late` _does_ exist in the grammar, but `late static` is always a syntax error, so we might have a bug where `static late` is accepted in this case because it is accepted in some cases, and that one is perhaps more likely than the bug where `late static` is accepted, anywhere.