Fixes #2355. Add more typed_date.setRange() tests (#2412)
This PR adds many new tests. The idea is to test each `setRange()` of an `int` type against a `float` type of the same size (to ensure that we have an error, not a memcpy). Next, we test it against a type of a smaller size, and against a type of a bigger size. We do the same for each `float` against `int`.

We have float types with sizes 32, 64, and 128 (32x4 and 64x2) only. Therefore we test `Int16` against `Uint8` and `Uint16` in case of small and equal size, which is not an error. 

When we are checking 32-bit type against 16, 32 and 64 bit, the latter case is moved to separate tests because it has to be switched off on web configurations.

Issue dart-lang/sdk#53945 exists for `x` types as well, so we test them, too.